// src/oncobridge-web/src/app/api/models/import-response.ts
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

import { SourceResourceResponse } from '../models/source-resource-response';
export interface ImportResponse {
  bundleType?: (string | null);
  contentHash: string;
  entryCount: (number | string);
  fileName?: (string | null);
  importBatchId: string;
  normalizedAt?: (string | null);
  normalizerVersion?: (string | null);
  patientIds: Array<string>;
  receivedAt: string;
  sourceResources: Array<SourceResourceResponse>;
  sourceSystemLabel: string;
  status: string;
}