// src/oncobridge-web/src/app/api/models/source-resource-response.ts
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

import { JsonElement } from '../models/json-element';
export interface SourceResourceResponse {
  contentHash?: (string | null);
  entryIndex: (number | string);
  fullUrl?: (string | null);
  id: string;
  resourceJson?: (null | JsonElement);
  resourceType?: (string | null);
  sourceLogicalId?: (string | null);
}